# reduce allocation by 15%